[Feat/#137] 스터디 그룹 멤버 추가 api연결 및 수정 사항

+++ data/src/main/java/com/umc/data/api/OrganizationApi.kt
@@ -167,7 +167,7 @@ interface OrganizationApi {
 
     @PUT(Endpoints.Organization.STUDY_MEMBER)
     suspend fun changeGroupMember(
-        @Path(PATH_GROUP_ID) groupId: Int,
+        @Path(PATH_GROUP_ID) groupId: Long,
         @Body request: ChallengerListRequest
     ): ApiResponse<Unit>
